graphical  upgrade of forecast function
It has also introduced a dependency for the package tidybayes, which helped to make a better legend thanks to the geom_lineribbon function. Probably also with base ggplot and its functions, geom_ribbon and geom_line, something similar could have been achieved, but it was much simpler to solve some graphical legend issues with this other function.